requirements.txt
langchain-community
langgraph
httpx
beautifulsoup4
jinja2
pyyaml
fastapi
uvicorn
